fix segv on cut

+++ srcs/line_editing/delete.c
@@ -5,7 +5,7 @@
 **	Moves every character in the string `shift_no' bytes to the left
 */
 
-void			shift_chars_left(char *str, unsigned int shift_no)
+void			shift_chars_left(char *str, size_t shift_no)
 {
 	size_t i;
